refactor: flex처리
