+++ client/src/components/map/useKeywordSearch.ts
@@ -13,7 +13,6 @@ const useKeywordSearch = (
   radius?: number
 ) => {
   const map = useSelector((state: RootState) => state.map.map);
-  const ps = new kakao.maps.services.Places();
   const dispatch = useDispatch();
 
   const placesSearchCB = (
@@ -35,6 +34,7 @@ const useKeywordSearch = (
   };
 
   useEffect(() => {
+    const ps = new kakao.maps.services.Places();
     ps.keywordSearch(searchPlace, placesSearchCB, {
       x: Number(x),
       y: Number(y),
